refactor(api): seperate resource phases
